moved queries in module
