+++ src/spatialdata/_core/operations/rasterize.py
@@ -294,9 +294,7 @@ def _get_xarray_data_to_rasterize(
             m = corrected_affine.inverse().matrix  # type: ignore[attr-defined]
             m_linear = m[:-1, :-1]
             m_translation = m[:-1, -1]
-            from spatialdata._core.operations.spatial_query import (
-                get_bounding_box_corners,
-            )
+            from spatialdata._core.query.spatial_query import get_bounding_box_corners
 
             bb_corners = get_bounding_box_corners(
                 min_coordinate=min_coordinate, max_coordinate=max_coordinate, axes=axes
