+++ src/spatialdata/_core/query/spatial_query.py
